@@ -1,6 +1,6 @@
 {
   "name": "remix-easy-mode",
-  "version": "0.1.1",
+  "version": "0.1.2",
   "author": {
     "name": "Sam Cook",
     "email": "email@samcook.cc",
@@ -15,26 +15,25 @@
     "dist"
   ],
   "dependencies": {
-    "@kiruna/form-data": "0.0.0-alpha.7",
-    "@kiruna/promises": "0.0.0-alpha.7"
+    "@kiruna/form-data": "0.0.0-alpha.31"
   },
   "devDependencies": {
-    "@remix-run/eslint-config": "^1.18.0",
-    "@remix-run/react": "^1.18.0",
-    "@remix-run/server-runtime": "^1.18.0",
-    "@types/react": "^18.0.33",
-    "eslint": "^8.38.0",
+    "@remix-run/eslint-config": "^2.2.0",
+    "@remix-run/react": "^2.2.0",
+    "@remix-run/server-runtime": "^2.2.0",
+    "@types/react": "^18.2.33",
+    "eslint": "^8.52.0",
     "react": "^18.2.0",
-    "rimraf": "^5.0.0",
-    "tsup": "^6.7.0",
-    "typescript": "^5.0.4",
-    "zod": "^3.21.4"
+    "rimraf": "^5.0.5",
+    "tsup": "^7.2.0",
+    "typescript": "^5.2.2",
+    "zod": "^3.22.4"
   },
   "peerDependencies": {
-    "@remix-run/react": "^1.18.0",
-    "@remix-run/server-runtime": "^1.18.0",
-    "react": "^18",
-    "zod": "^3"
+    "@remix-run/react": ">=1.18.0",
+    "@remix-run/server-runtime": ">=1.18.0",
+    "react": ">=18",
+    "zod": ">=3"
   },
   "scripts": {
     "build": "rimraf dist && tsup index.ts --format cjs,esm --dts",
