@@ -1,7 +1,7 @@
 import React from "react"
 import { InputHelper } from "../../../../src/unstyled-components/input-helper"
 import { useExampleHook } from "./api"
-import { TextInput } from "@mantine/core"
+import { TextInput, MantineProvider } from "@mantine/core"
 
 const Comp = React.forwardRef<
   HTMLInputElement,
@@ -15,69 +15,71 @@ export default function Index() {
 
   return (
     <div>
-      <Form
-        csrfToken="5"
-        onSuccess={(successRes) => console.log("form onSuccess", successRes)}
-        style={{ display: "flex", flexDirection: "column", gap: "1rem" }}
-      >
-        <label>
-          Any old string (do not type "bad message")
+      <MantineProvider>
+        <Form
+          csrfToken="5"
+          onSuccess={(successRes) => console.log("form onSuccess", successRes)}
+          style={{ display: "flex", flexDirection: "column", gap: "1rem" }}
+        >
+          <label>
+            Any old string (do not type "bad message")
+            <InputHelper
+              {...fields.anyString.props}
+              defaultValue={"any old string"}
+              component={TextInput}
+              errorProps={{
+                error: fields.anyString.errors?.[0]?.message,
+              }}
+            />
+          </label>
+
+          <label>
+            Hello world (literal)
+            <InputHelper
+              {...fields.helloWorld.props}
+              defaultValue={"hello world"}
+            />
+          </label>
+
+          {/* ZOD UNION OF STRING LITERALS */}
+          {fields.letters.options?.map((option) => {
+            return (
+              <label key={option}>
+                {option}
+                <InputHelper
+                  type="radio"
+                  {...fields.letters.props}
+                  value={option}
+                  defaultChecked={option === 2}
+                />
+              </label>
+            )
+          })}
+
+          {/* ZOD ENUM */}
+          {fields.letters2.options?.map((option) => {
+            return (
+              <label key={option}>
+                {option}
+                <InputHelper
+                  type="radio"
+                  {...fields.letters2.props}
+                  value={option}
+                  defaultChecked={option === "1"}
+                />
+              </label>
+            )
+          })}
+
           <InputHelper
-            {...fields.anyString.props}
-            defaultValue={"any old string"}
-            component={TextInput}
-            errorProps={{
-              error: fields.anyString.errors?.[0]?.message,
-            }}
+            type="number"
+            {...fields.someNumber.props}
+            defaultValue={0}
           />
-        </label>
 
-        <label>
-          Hello world (literal)
-          <InputHelper
-            {...fields.helloWorld.props}
-            defaultValue={"hello world"}
-          />
-        </label>
-
-        {/* ZOD UNION OF STRING LITERALS */}
-        {fields.letters.options?.map((option) => {
-          return (
-            <label key={option}>
-              {option}
-              <InputHelper
-                type="radio"
-                {...fields.letters.props}
-                value={option}
-                defaultChecked={option === 2}
-              />
-            </label>
-          )
-        })}
-
-        {/* ZOD ENUM */}
-        {fields.letters2.options?.map((option) => {
-          return (
-            <label key={option}>
-              {option}
-              <InputHelper
-                type="radio"
-                {...fields.letters2.props}
-                value={option}
-                defaultChecked={option === "1"}
-              />
-            </label>
-          )
-        })}
-
-        <InputHelper
-          type="number"
-          {...fields.someNumber.props}
-          defaultValue={0}
-        />
-
-        <button type="submit">Submit</button>
-      </Form>
+          <button type="submit">Submit</button>
+        </Form>
+      </MantineProvider>
 
       <pre>{JSON.stringify(result, null, 2)}</pre>
     </div>
